Modulase the code
